EntitySession#last_persisted_seq
To be used for optimistic locking when appending new events to store.
We tell the store what’s the last event sequence we know about for a session.

It is initialized to #seq when loading, and only updated when clearing events, so we can potentially keep working with a session instance and persist again.